Added TraceSubscription to Tracer interface and re-worked the interface
This is a breaking change, but I think the previous interface had flawed like limited
support for root operation, name was not totally right since `TraceQuery` was picking
up `mutation` change also and no support to make some analysis on the selection of field.

The new interface makes a distinction between a Request and actual root "operation". In a
request, the new tracing format supports the following nesting:

Request (query)
 Root query (type, alias)
   Field1
   Field2
 Root query (type, alias)
   Field1
   Field2
End

The `TraceRequest` could be renamed `TraceOperation` with a specific kind, but `TraceRequest`
felt.

Dependent on having the actual selection of fields.